Mesh page - Add column linking remote Kialis
This adds a column in "Mesh" page to show the Kiali instances discovered in each listed cluster.

Format of the column is "namespace / service name". If the Kiali service has the "kiali.io/external-url" annotation, the value of that annotation is used to create links and allow the user to jump to those remote Kialis.

Fixes kiali/kiali#3526