Fix the traffic animation issue (#2108)
* Fix the traffic animation issue, which turned out to just be a symptom of
a deeper problem.  It happened only when animation was enabled and the
user enabled boxing, changed to an "app" graph type, or navigated to a
graph with boxing (and yes, this took a long time to realize :)). The
root cause was the fact that boxLayouts finish in an async fashion, and
animation was started prior to the layout completing. Theoretically it
could also happen without any boxing, but just using cola or cose-bilkent
layouts, which can also run async.

The solution is to introduce some promise logic to ensure we don't complete
a graph update before a layout is complete.

* unrelated but included, a small change to legend labels based on some
  confusion reported by a community member.

https://github.com/kiali/kiali/issues/3755