Apply UX feedback on graph boxing, and update PF color application (#2117)
Apply UX feedback regarding the recent MC-related additions of namespace and cluster boxing. Specifically:
- restrict box colors to the [light] gray PF palette for uniformity and in order to accentuate status-oriented color use in the graph
- darken the namespace and cluster box borders to make them more distinct
- add PF badging on box labels to make them easier to locate
- and increase font size on non-traffic box labels (i.e. leave app boxes as is for better uniformity)
- reduce highlighting "flashing" by not highlighting/dimming the contents of namespace and cluster boxes. in general, users may just be navigating over these boxes to reach a more specific node
  - i.e. only highlight the box labels for non-traffic boxes

While working on the above changes I noticed several issues in our "PfColors.tsx":
- colors were hardcoded
- many colors were out of date as the values were set using PF3
- many defined colors were unused
- the color definitions were somewhat haphazard

I've overhauled PfColors.tsx to resolve the issues. With the changes:
- we now define only the colors we use
- all colors are now resolved via PF variables such that any changes in PF will be picked up moving forward
  - note: we do still have a couple of custom colors defined by UX
- there are now two specific ways to use the defined colors, the common CSS mechanism, and another for when the specific value must be resolved (for cytoscape use).

So, with the conversion to the PF variables some colors have been updated. In general the changes will likely not be noticed, but now we comply with PF4 (at least the version we are using).

The bulk of the PR "meat" is in GraphStyles.ts, PfColors.tsx, and GraphHighlighter.tsx. Most everything else is just a touch due to the color refactor.

https://github.com/kiali/kiali/issues/3675