Navigation to remote Kialis from graph context menu and side panel (#2118)
* Add navigation to remote Kiali in context menu and side panel

This changes graph context menus for nodes that represent resources in non-local clusters. The context menu was trying to open a URL in the local Kiali, but the right action is to open a "remote Kiali instance" (if any) where the resource can be inspected. It's the same case for side panel kebab menu and main links.

When building the graph node context menu, options are made available only if the node belongs to the "local" cluster, or if the node belongs to a remote cluster where a Kiali was successfully discovered and is accessible. Otherwise, no options in the context menu are made available.

Icons are used to denote external links in graph context menu and side panel If remote Kiali is unreachable, the context menu will be empty. Also, the kebab menu in the side panel will be empty. I case of textual link (main links in side panel) a plain text is shown rather than links.