Terminate gracefully on errors in OpenID's authorization code flow
Although the OpenIdCodeFlowHandler was terminating using the "RespondWithError" or "RespondWithDetailedError" helper functions, these functions return a JSON under the assuption that the Kiali UI did the request.

However, the OpenIdCodeFlowHandler is not called by the Kiali UI, but by the browser directly. This was causing a bad experience, because the user would see the returned JSON in the browser window.

These changes change this behavior: errors are logged and, additionally, a redirection to boot the Kiali UI is done. The redirection includes an "openid_error" parameter to let know the UI that there was an error and show the message in the login page.

Fixes #3574