Friendly error messages for OpenID
The UI wasn't showing the error messages passed in the URL by the OpenID provider. This make sure error messages from the IdP are displayed in the login page.

Also, when Kiali is using the authorization code flow, any error was being returned as a bare JSON in the browser. But back-end is having a behavioral change and will be doing a redirection with the error in an "openid-error" parameter. These changes will also show the error returned by kiali back-end in the login form.

Related kiali/kiali#3574