Always use a locally built kiali-ui in get-console.sh
Since the kiali-ui will no longer be published to NPM, makefiles should no longer rely on NPM to get a console. It's now assumed that developers should checkout the kiali-ui source code and build it.

Additionally, this is removing usage of both the `jq` and `npm` commands in get-console.sh. Instead, the `sed` command is used, which is usually present in Linux distributions.

Related #3891