Refine message for remote Kiali (#2137)
* Refine message for remote Kiali

This standardizes the behavior of the graph context menu when there are no available options to show. Concretely, no options because of a node living in a remote cluster where an accessible Kiali instance couldn't be found.

On Kiali first load, it will check which clusters does NOT have accessible Kialis. A message will be logged in the message center telling about this fact and that no links to these clusters will be available. Context menus for relevant nodes won't be present.

Fixes kiali/kiali#3907