Fix mini-graph update after detail page Actions (#2142)
* Fix mini-graph update after detail page Actions
- also maintains previous fix for update after mini-graph navigation

https://github.com/kiali/kiali/issues/3983

* Fix double-fetch issues in the detail Info pages.  The primary problem
was due to both the Detail and Info components checking for changes to
"lastRefreshAt".  The Info component should defer this check to the Detail
component.  On a refresh the Detail component already re-fetches the
target object (app, workload, service), when the re-fetch completes it
will then cause the Info component to update and re-fetch the mini-graph.

Note that the Object equality comparisons in the Info componentDidUpdate
impls  are appropriate, as the compare-by-reference works fine given that
the a new target object is produced on each re-fetch performed by the
Detail component.

Finally, ensure each of the Info components defers a graph update until it
has a valid target element.