Allow read-only file system in Kiali's pod
Specifically, stop writing changes to the index.html and env.js files, which come from the UI package.

These files are now generated on the back-end and served directly, rather than changing the files on the pod filesystem.

This allows using `readOnlyRootFilesystem: true` for the securityContext in the Kiali pod.

Fixes #3810