Deprecate publishing to NPM (back-end) (#3986)
* Always use a locally built kiali-ui in get-console.sh

Since the kiali-ui will no longer be published to NPM, makefiles should no longer rely on NPM to get a console. It's now assumed that developers should checkout the kiali-ui source code and build it.

Additionally, this is removing usage of both the `jq` and `npm` commands in get-console.sh. Instead, the `sed` command is used, which is usually present in Linux distributions.

* Adjust run-kiali.sh to not download kiali-ui from NPM
* Changes to Jenkinsfile to no longer publish to NPM
  Also, remove the determine-release-type.sh file. This script is now embedded in the Jenkinsfile.
  Some changes to deploy/jenkins-ci/README.md were done, but it's not ready yet.

* Update kiali-release job for the testing Jenkins

Update the configuration of the kiali-release job that is setup for the Jenkins image that is part of the developer setup to try the release pipeline.

* Update the README.adoc file
  * Remove outdated parts
  * Reorganize so that a development setup can be achieved by a top-bottom read rather than having to jump to different sections as needed.

* Update releasing docs
* Update Jenkin's README.md
* Update main RELEASING.adoc

Co-authored-by: John Mazzitelli <mazz@redhat.com>

Related #3891