Use a mounted certificate, rather than a ca_file config
Remove the recently added ca_file option from the config. Instead, look for a file in the path "/kiali-cabundle/openid-server-ca.crt". If it exists, it's assumed it is the CA certificate for the OpenID server. It's used on connections to the openid server.

The Operator is going to configure the Kiali Deployment to properly mount this file, if certain configmap is present.