Support for custom CA certificate for OpenID auth (#4112)
* Support for custom CA certificate for OpenID auth

Adding support for custom CA certificate for the OpenID authentication.

The CA certificate is provided through a volume mount. Thus, this checks for existence of the certificate file and uses it, if available.

The Operator is going to configure the Kiali Deployment to properly mount this file, if certain configmap is present.