Run molecule tests with KinD (#4189) (#4233)
* run molecule tests with kind

* Use volumes rather than mounts, and use full fork names

* Prefer using volumes rathern than mounting directories on the host
machine into the container. Although this was mainly to add
compatibility for using a remote docker engine, in general, this
potentially works better. The rationale is that there is no need to
adjust privileges nor SELinux on the host for the container to work
properly, because the container engine is fully managing the volume.
* Use forks's fullname rather than just the username/orgname that has
the fork. This helps with older forks that have the old swscore name.

* Use Libera IRC instead of Freenode. Also add KinD config

Since Kiali has move to Lbiera Chat for IRC, use it instead of Freenode.
Also add a configuration for KinD to start with a control plane node and
a worker node.

Co-authored-by: John Mazzitelli <mazz@redhat.com>

Co-authored-by: Edgar Hernández <ehernand@redhat.com>