Add new gateway_hostnames graph appender (#4138)
* Add new gateway_hostnames graph appender

This new appender will find Ingress Gateway nodes (workloads) in the graph and flag them appropriately. It will also find any associated Istio Gateway CRD that configure certain Gateway in order to attach to the node the list of hostnames that the Gateway is serving requests to.

A similar logic was added to also provide hostnames for VirtualServices.

The new "appender" or new logic is integrated into the already existent istio_details appender.