Add hostnames list for gateway and virtual service nodes (#2196)
* Add hostnames list for gateway nodes

In the main graph, when clicking on nodes representing ingress gateways, the side panel will provide the list of configured hostnames for the selected gateway. This is shown in an expandable section. In a similar way, host for VirtualServices are aso shown in the side panel.

In the graph, hosts are also shown. The graph will, initially, show the text "X hosts" on nodes that represent Ingress Gateways or nodes that have attached a VirtualService. Hovering the mouse's cursor over this text will expand the text to show what are those hosts limited to showing up to 5 hosts. If the node has associated more than 5 hosts, a legend stating this fact will appear at the end of the list.

Also, there is a new badge to mark/identify graph nodes that are detected as ingress gateways.