Reduce logging related to clustering discovery
When trying to discover remote clusters, some secrets are queried from the Istio namespace. If Kiali has got some reduced privileges, a warning message telling about forbidden permissions is repeatedly logged, which causes some unneeded noise in the logs.

This is detecting the forbidden response and replacing the warning-level message in the logs with a trace-level message, hopefully reducing the noise in the logs.

Related kiali/kiali#4147.