Add session persistor
The session persistor will do the work of creating, reading and deleting session cookies. It is based on how the OpenID -> authorization code flow is doing session persistence.

Right now, the cookie handling logic is spread in the code of the various strategies. It is somewhat duplicated. This is to reduce code duplication and as part of the removal of the dgrijalva/jwt library, because session data storage is the main client of this JWT library.