Remove usage of the dgrijalva/jwt library for the Token authentication (#4596)
* Add session persistor

The session persistor will do the work of creating, reading and deleting session cookies. It is based on how the OpenID -> authorization code flow is doing session persistence.

Until now, the cookie handling logic was spread in the code of the various strategies. It is somewhat duplicated. This is to reduce code duplication and as part of the removal of the dgrijalva/jwt library, because session data storage is the main client of this JWT library.

* Add handler for Token authentication as an isolated controller

This is a working version of Token authentication without JWT
Token authentication seems to be working OK now without using the dgrijalva/JWT library.
Because of the new unit tests, remove other tests that were superseeded