Add Finalizers to the graph generation architecture (#4731)
This is pure code organization at the moment.  But I think it may be useful for some of the
projected features.  I am tying this to a current epic, which likely
will want to add k8s label info to graph nodes. This could be done in
an appender, but may lend itself better to being done one time, at the end.
Even if we don't add more finalizers, it's still not a bad thing.  There
isn't really a perf hit and I think it adds some structure.

https://github.com/kiali/kiali/issues/4605