Fix Prometheus TSDB retention config (#4785)
Prometheus deprecated "storage.tsdb.retention" many versions back, and
it is no longer valid.  Unfortunately it is still set, but to 0, invalidating
this information in serverConfig.  Switch to "storage.tsdb.retention.time",
the valid setting.  Also, add a check to ensure this setting is never 0.
Furthermore, ensure that if all fails we provide a valid default to
consumers of the config.

https://github.com/kiali/kiali/issues/4784