Clean-up funcs and types related to JWT generation
In config/token.go, remove GetSignedTokenString and GenerateToken. These functions created a token for the very old "login" strategy in which users logged in to Kiali by typing a username and a password in the login form. It was possible to configure a single username/password combination (no multi-user support). This auth strategy was removed long ago, but these funcs remained in place to support some unit tests.

In consequence, the TokenGenerated type can be removed, because it is no longer used which supported usage of JWTs for Kiali's session handling.

Unit tests are fixed by no longer injecting a stale token when testing authentication requests. Those situations should be already covered in session_persistor_test.go units.