Do not try to find session info for Anonymous auth strategy
When "anonymous" auth strategy is enabled, it is not expected that any cookie nor HTTP Authorization headers will be populated. So, the AuthenticationInfo func should not try to search for a session in this case. This caused the removal of some funcs that became unused.

Additionally, the TokenResponse type was removed. It is not used anymore.