Deprecate usage of deleteTokenCookies func
In the Logout func it does not make sense to call it, because no cookies are expected to be present/set by Kiali. In other parts of the code, its replacement is the TerminateSession func of the active "authentication controller" (look at files under the business/authentication path).

Also, remove some IDE warnings (variable naming collisions)