Fix Jenkins releases
Jenkins was still using the kiali/kiali-ui repo and bundling that. These
changes should update Jenkins to only use the unified kiali/kiali repo.
This is a simple "patch" since it is know that we are already in the
process of phasing out Jenkins (i.e. this is just best effort patches).