Remove dgrijalva/jwt dependency: refactor Header auth and auth-related code clean up (#4867)
* Integrate *api.AuthInfo into the UserSessionData type
* Rename "sessionPayload" to "tokenSessionPayload", because each auth controller ended up having it's own payload type.
* Move UserSessionData and AuthenticationFailureError types from token_auth_controller.go to auth_controller.go, because these types ended being shared types.
* Clean-up funcs and types related to JWT generation

In config/token.go, remove GetSignedTokenString and GenerateToken. These functions created a token for the very old "login" strategy in which users logged in to Kiali by typing a username and a password in the login form. It was possible to configure a single username/password combination (no multi-user support). This auth strategy was removed long ago, but these funcs remained in place to support some unit tests.

In consequence, the TokenGenerated type can be removed, because it is no longer used which supported usage of JWTs for Kiali's session handling.

Unit tests are fixed by no longer injecting a stale token when testing authentication requests. Those situations should be already covered in session_persistor_test.go units.

* Do not try to find session info for Anonymous auth strategy

When "anonymous" auth strategy is enabled, it is not expected that any cookie nor HTTP Authorization headers will be populated. So, the AuthenticationInfo func should not try to search for a session in this case. This caused the removal of some funcs that became unused.

Additionally, the TokenResponse type was removed. It is not used anymore.

* Deprecate usage of deleteTokenCookies func

In the Logout func it does not make sense to call it, because no cookies are expected to be present/set by Kiali. In other parts of the code, its replacement is the TerminateSession func of the active "authentication controller" (look at files under the business/authentication path).

Also, remove some IDE warnings (variable naming collisions)

* More cleanup: Constant usage unification and cleanup

Removing unused constants. Also, using AESSessionCookieName rather than TokenCookieName+"aes".

* Remove the dgrijalva/jwt-go library

* Replace it with the gopkg.in/square/go-jose.v2/jwt library in the only remaining usage.
* Run `go mod tidy`
* Remove its mention in the style guide.

* Fix swagger doc

* Fix argument in call: it should be passed as pointer

* Remove token check in ValidateSession

Original author of the header auth reports that the equality check between the token found in HTTP Headers and the one found in cookies will break things.

* Feedback: Add compatibility with Cookies Version 0

This patch is provided by @mlbiam.