Remove KIA1003 and KIA1103 checkers
These checkers are now obsolete, because Istio is actively ensuring that the offending configurations verified by KIA1003 and KIA1103 cannot make it into the cluster (i.e. the cluster will reject a resource with the offending configs). This means that KIA1003 and KIA1103 are now dead checkers and can be removed safely.