Update confirmation button placement and styling (#5124)
* The initial PF update left the confirmation box buttons in an intermediate
state.  PF styling moved the buttons from bottom-right to bottom-left,
and as such the button order needed to change such that "Cancel" is the
inner, not the leading, button.  Furthermore PF styling calls for the
"Cancel" button to be the "link" variant.

See https://www.patternfly.org/v4/components/modal/design-guidelines/#confirmation-dialogs

* PR Feedback (lucas, andrew):
- to better align with console, use 'secondary' as opposed to 'link' for "close" buttons

Also:
- consistently use PF constants for button variant settings

* fix linter issue

* add one more file change

* another small fix