Move to v1/CronJob Api
This is because the v1beta1/CronJob Api is going to be removed in the next Kubernetes release. Also, since Istio 1.14 no longer supports Kubernetes versions prior to v1.20, it no longer makes sense to support the older Api in Kiali.

OpenShift 4.9+ has Kubernetes API 1.22+. So, also on this platform is safe to leave behind the older api.

 This removes warning messages from Kiali logs.

 Fixes #4519.