Stream pod logs while processing
The pod logs endpoint was loading the pod logs on memory, then parsing/process them (also on memory) and, finally sending the results to the browser/client. This is leading to high memory usage for large logs.

This is changing the approach to parse/process logs and send results to the client/browser while also reading, reducing memory usage to a minimum.