Restore getParsedLogs in business/workloads.go
It is not possible to stream logs when a Duration (endTime) and TailLines parameters are set. This is because the cluster API does not support end time parameter. So, Kiali needs to load some logs into memory to properly resolve the lines that are in the tail of the requested time range. In consequence, it is not possible to send results until the resultset is resolved.

However, since the number of tailLines is known it is possible to limit the amount of memory used, rather than use an indefinite amount. So, some changes to getParsedLogs were done to keep memory usage limited.

Additionally, a bug was solved when SinceTime and Duration params are both set: depending on available logs, it was possible that the returned logs didn't belong to the requeted time window.