Back-end: Allow returning metrics for both reporters
This removes the constraint that a reporter is required in the back-end for querying metrics. To return metrics for both reporters, the URL argument should be reporter=both which will generate a prometheus query without the reporter label. This would, of course, aggregate values for both reporters in a single timeserie. In order to have them properly separated, the client should include a byLbl[]=reporter parameter in the URL, which is already an existent/supported parameter.