Launch wizards from graph's context menu
At first, GraphPage.tsx was planned to store the service details on its state to share it between the side panel and the graph context menu. However, the context menu can be open for a different node than the one for the side panel which means that sharing data does not make sense. As a consequence, some code and variables were pushed down on the component tree. The GraphPage.tsx is now limited to only showing the wizard or showing the delete confirm dialog.