Show time controls in a modal from mini-graph card
**This only affects the OpenShift plugin**

The _refresh interval_ and _duration_ controls are currently not available in the OpenShift plugin, because toolbars are not visible.

When Kiali is running under the OS plugin context, the title of the mini-graph card (i.e. the time range) will become clickable. When clicking, a modal dialog will appear offering the missing time controls.

Since this is done in the mini-graph card component, these changes are reflected in all detail pages (apps, workloads, services).

Related to kiali/openshift-servicemesh-plugin#42