Show time controls in a modal from mini-graph card (#5389)
* Show time controls in a modal from mini-graph card

**This only affects the kiosk=parent mode**

The _refresh interval_ and _duration_ controls are currently not available in the kiosk mode, because toolbars are not visible.

When Kiali is running under in kiosk mode, the title of the mini-graph card (i.e. the time range) will become clickable. When clicking, a modal dialog will appear offering the missing time controls.

Since this is done in the mini-graph card component, these changes are reflected in all detail pages (apps, workloads, services).

Related to kiali/openshift-servicemesh-plugin#42