Fix test validating that the degraded status of a service is reported (#5454)
This is a cypress tests which was failing at times. Looks like, the
source of the issue is the environment which sometimes records more
errors than expected leading to an error status of a service that is
expected to give only up to degraded status.

This changes Kiali configuration to extremely tolerant with the relevant
service: only if 100% of the traffic is erring, the service will be
flagged as faulty. Otherwise, its health will be degraded as long as
more than 2% of the traffic fails.

Fixes #5226