Remove the prometheus-performed rounding on metrics queries. (#5466)
Remove the prometheus-performed rounding on metrics queries.  The OR-based
query requires prometheus to basically query the data twice, once with rounding
and once without.  The data may still have raw numbers, and so it's
questionable whether the rounding provides much value.  It does return
a more compact response, but the metrics responses are not particularly
large, so the bytes-savings is not worth the extra load on prom or the
extra time to generate the response.  Clients can perform any required
rounding of the results.

- fix sparkline tooltips to show 2 fixed decimal points (thanks lberetta)

https://github.com/kiali/kiali/issues/5432