Do not omit namespace when querying for traces
In the past, the name recorded for services in the istio namespace didn't include the namespaces in traces. Currently, that is no longer the case, apparently.

Since Kiali was omitting the namespace for services in the istio namespace, traces/spans of such services weren't being fetched and were missing in Kiali. Most notably, this was the case for istio gateways workloads (either ingress or egress).

This is removing the conditional to omit the namespace based only on configuration flags.

Fixes #4935