Add Control Plane status indicator in Overview card
This replicates the status indicator for control plane components in the control plane card of OverviewPage.tsx.

For the card, the indicator icon is changed to match the icons used in the tooltip. The masthead indicator is unchanged.

The shown/hidden behavior of the indicator in the control plane card is the same as in the masthead: it is hidden if all components are OK, and it is shown if something has some potential issue.

Fixes #5525