It seems there is a legitimate case (possibly with a newer version of (#5699)
Istio, I'm not sure) where the telemetry may have app set, but not
workload or service.  Possibly when the workload is not injected but
app/version labeling is still applied to the workload.

In this case, avoid a graph-gen panic when trying to generate a workload
graph node, and instead default to an Unknown service node.  Panics don't
really help anyone, so anything we can do to avoid that situation is, I
think, better.

https://github.com/kiali/kiali/issues/5696