Fetch OpenShift version using Kiali ServiceAccount
This change applies only to the OpenShift authentication mechanism.

Previously, the cluster version was retrieved using an unauthenticated query to the cluster API. This is ensuring that such query is using authenticated using the Kiali ServiceAccount. This is to overcome situations when non-authenticated users cannot query the cluster version.

Fixes #5755