[v1.65] Kiali does not handle namespace-level injection enable/disable for Maistra  (#6175)
Kiali does not handle namespace-level injection enable/disable for Maistra. (#6172)
Update Cypress to handle that use-case.
OSSM-3989

I'm unable on my CRC to reproduce a button click failure seen in downstream.  
Adding 'be.enabled' button requirement to see if it makes a difference (I doubt it but we can try...)