An approach to handling the "mutiple badge" issue for Nodes (#6392)
Because PFT allows for only a single labelIcon (i.e badge), we will reserve that specifically to indicate Traffic Sources (or Gateways) because it helps users identify where traffic is entering the mesh.  The other badges we see in Cytoscape will be presented as node attachments.  BUT, since there are only a limited number of attachments (4), and we want to present badges in a consistent location, all VS features will be indicated by the VS icon, and the tooltip will indicate the specific VS features in play.

Also:
- reduce node size to make Mazz happier

https://github.com/kiali/kiali/issues/6370