Fix regression introduced in https://github.com/kiali/kiali/pull/6087.  (#6403)
The 6087 PR successfully fixed an issue with caused by premature rendering of the SummaryPanel.  But the overly-delayed rendering caused a resize of the available graph space, which forced a unwanted "fit".  This PR changes the approach, delaying an appropriate amount.  There is still a small window of vulnerability, but in general users should not see a problem.

https://github.com/kiali/kiali/issues/6248