Refactor how Urls for tracing tools are generated
Change moves all the URL generation code into tracing URLProviders,
which are classes that provide specific URL generation methods for each
Tracing frontend supported (Jaeger, Grafana). This change removes all
the templating strings that where previously in a number of different
locations, and makes it much easier to provide feature parity between
Jaeger and Grafana.

More than that it makes it much easier to introduce support for new
tracing frontends, as only a single new class needs to be created for
that frontent, and the rest of the plumbing is done for free. This
modular approach to tracing frontends is particularly useful for
supporting tracing frontends with complex URL schemas, such as Grafana
v10 and higher.

The modular approach also makes it possible for backend like Tempo to
declare themselves as a frontend provider, then silently forward URL
generation requests to whatever frontend is actually setup. For Tempo it
creates the possibility of using Kiali with Tempo, and a Jaeger frontend
instead of Grafana.