Add support for Grafana 10+ URLs
Grafana 10.1+ onwards (and probably also 10.0, the docs aren't clear)
has removed the URL schema that explore used in before 10.0, and
replaced it with a new documented schema for explore. This change adds
frontend support for that new schema, but is still missing the needed
backend support to transport new config values needed for URL
generation.