Backend changes to support new tracing URL generation for Grafana 10+
Grafana 10+ introduces a new url schema that requires additional
configuration information to produce working URLs. This change add those
additional config parameters to the backend tracing config, and plumbs
them into the Istio Status API, so they can be synced to the frontend
and use in URL generation.

As an initial first pass, this change isn't the cleanest, and is lacking
in comments etc. Will be added later once it's agreed that this approach
is correct.