Add support for new Grafana 10+ URL schema (#7085)
* Refactor how Urls for tracing tools are generated

Change moves all the URL generation code into tracing URLProviders,
which are classes that provide specific URL generation methods for each
Tracing frontend supported (Jaeger, Grafana). This change removes all
the templating strings that where previously in a number of different
locations, and makes it much easier to provide feature parity between
Jaeger and Grafana.

More than that it makes it much easier to introduce support for new
tracing frontends, as only a single new class needs to be created for
that frontent, and the rest of the plumbing is done for free. This
modular approach to tracing frontends is particularly useful for
supporting tracing frontends with complex URL schemas, such as Grafana
v10 and higher.

The modular approach also makes it possible for backend like Tempo to
declare themselves as a frontend provider, then silently forward URL
generation requests to whatever frontend is actually setup. For Tempo it
creates the possibility of using Kiali with Tempo, and a Jaeger frontend
instead of Grafana.

* Add support for Grafana 10+ URLs

Grafana 10.1+ onwards (and probably also 10.0, the docs aren't clear)
has removed the URL schema that explore used in before 10.0, and
replaced it with a new documented schema for explore. This change adds
frontend support for that new schema, but is still missing the needed
backend support to transport new config values needed for URL
generation.

* Replace strings with constants

Grafana, Jaeger and Tempo are all string used in multiple places,
swapping them all out for constants defined in a single location.

* Backend changes to support new tracing URL generation for Grafana 10+

Grafana 10+ introduces a new url schema that requires additional
configuration information to produce working URLs. This change add those
additional config parameters to the backend tracing config, and plumbs
them into the Istio Status API, so they can be synced to the frontend
and use in URL generation.

As an initial first pass, this change isn't the cleanest, and is lacking
in comments etc. Will be added later once it's agreed that this approach
is correct.

* Update Tempo specific config. Fix for TraceID

---------

Co-authored-by: Thomas Purchas <purchas@apple.com>
Co-authored-by: josunect <josune.cordoba@gmail.com>