Note to self...
This commit has pretty decent control of external deployments, now
in a dedicated, special, cluster box outside of any other cluster box.

It also eliminates "dataplanes" naming, which made no sense if favor of
"Data Plane" naming. A Data Plane node now has
 secondary labeling for the namespace count.

TODO! The TargetPanelDataPlane stuff is still a mess, it's partially
implemented to show an expandable list of namespaces. The idea is to
show the NamespaceCard info on Expand, but it's not in good shape, also
I think it's fetching them all up front, which we don't want to do.