[pft graph] handle parallel edges (#7419)
In the early days of the PFT graph dev we overrode PFT's bendpoints to
force straight edges, like we have in Cytoscape.  But in the case that
the same source and dest nodes (a -> b) has parallel edges (typically
edges with different protocols), the removal of bendpoints caused the
edges to both be straight, thus overlaying. This adds a check to allow
bendpoints in this situation.

Note - as PFT improves its layouts we may want to revisit this customization.