Fix issue where we mistakenly discarded http edges when tcp edges weren't also in the graph (and therefore no edges to/from waypoints).
The fix could have been centered around fixing the initialization of

  graphEdge := []*graph.Edge{}

To preserve non-tcp edges.  But the code was a little confusing to me, so
this reworks things a bit to [hopefully] make it more simple.

Note that there is probably still work to do, as TCP edge visualization is
still confusing.  If you add in waypoints you can see them but it's
unidirectional.  If you remove service nodes you mostly lose TCP
visualization.  But this should at least solve the current bug.