If a namespace appears to be deleted, stop the cache. This stops flooding the logs with errors that come from the k8s client cache. The kube_cache now calls back into kiali cache to clean up namespace stuff on namespace deletion. Kiali cache then loops back and calls the kube_cache so it stops the namespace informers.
This also restart the cache if a namespace is re-created. If the cache lister doesn't exist,
it is likely because the namespace was deleted and the informers were stopped;
if we detect that, the namespace was probably re-created so this also refreshes the cache. It uses a lock
to void deadlocks when getting listers concurrently.